chore(): separa a responsabilidade da camada service
